勝敗判定の実装 close #18, #19
